Remove legacy code and refactor
